Move test_binary_op_large_argument_list test to scalar_function module
